test(chat-form): add some tests for. the controls
